Let downloader cache a cookie in config.yml
The expiration date on the cookies returned by the TopCoder
API is measured in years. There's no reason to ask for a
new one every time you invoke the gettc command. This patch
will store the token in the config.yml file, and load it up
at init. Greatly reduces the number of "API limit exceeded"
errors you run into.